Test if two GADSdat objects are (nearly) equal

Run tests to check whether two GADSdat objects are (nearly) equal. Variable names, number of rows in the data, meta data and data differences are checked and reported as a list output.

Usage

equalGADS(
  target,
  current,
  id = NULL,
  metaExceptions = c("display_width", "labeled"),
  tolerance = sqrt(.Machine$double.eps)
)

Arguments

target

A GADSdat object.

current

A GADSdat object.

id

A character vector of length 1 containing the unique identifier column of both GADSdat. If specified, both GADSdat are ordered according to ID before comparing their data.

metaExceptions

Should certain meta data columns be excluded from the comparison?

tolerance

A numeric value greater than or equal to 0. Differences smaller than tolerance are not reported. The default value is close to 1.5e-8.

Value

Returns a list.

Details

More detailed checks for individual variables can be performed via inspectDifferences and inspectMetaDifferences.