eatGADS: Data management of hierarchical SPSS files via R and SQLite

The eatGADS package provides various functionality, mainly: importing data, data and meta data cleaning, creating a fixed form SQLite data base and using the SQLite data base.

Importing data

SPSS data (.sav) can be imported via import_spss. Further import functions exist as well: import_stata for importing Stata data (.dta), import_DF for importing R data.frames, import_RDS for importing R data.frames saved as .RDS files, and import_raw as well as import_raw2 for importing data from raw data and meta data files.

Data and meta data cleaning

Data cleaning functions include functions for recoding data (e.g., recodeGADS) or re-ordering variables (e.g., relocateVariable). Meta data cleaning functions include functions for changing variables labels (e.g., changeVarLabels), changing value labels changeValLabels or modifying missings tags changeMissings.

Creating a GADS data base

Hierarchical data sets are combined via mergeLabels and the data base is created via createGADS. For this, the package eatDB is utilized. See also createDB.

Using the GADS

The content of a data base can be obtained via namesGADS. Data is extracted from the data base via getGADS for a single GADS and via getTrendGADS for trend analysis. The resulting object is a GADSdat object. Meta data can be extracted via extractMeta, either from the GADSdat object or directly from the data base. Data can be extracted from the GADSdat object via extractData.

See also

Author

Maintainer: Benjamin Becker b.becker@iqb.hu-berlin.de

Other contributors:

Karoline Sachse [contributor]

Johanna Busse [contributor]