Inspect differences in a variable.

Inspect differences between two GADSdat objects in a specific variable.

Usage

inspectDifferences(varName, GADSdat1, GADSdat2, id)

Arguments

varName

A character vector of length 1 containing the variable name.

GADSdat1

A GADSdat object.

GADSdat2

A GADSdat object.

id

A character vector of length 1 containing the unique identifier column of both GADSdat.

Value

A list.

Details

Two GADSdat objects can be compared using equalGADS. If differences in the data are indicated for specific variables, these variables can be further inspected using inspectDifferences.

Examples

# create a second GADS with different data
pisa2 <- pisa
pisa2$dat$age[400:nrow(pisa$dat)] <- sample(pisa2$dat$age[400:nrow(pisa$dat)])

# inspect via equalGADS()
equalGADS(pisa, pisa2)
#> $names_not_in_1
#> character(0)
#> 
#> $names_not_in_2
#> character(0)
#> 
#> $data_nrow
#> [1] "all.equal"
#> 
#> $data_differences
#> [1] "age"
#> 
#> $meta_data_differences
#> character(0)
#> 

# inspect via inspectDifferences()
inspectDifferences("age", GADSdat1 = pisa, GADSdat2 = pisa2, id = "idstud")
#> $cross_table
#>         GADSdat2
#> GADSdat1 13.92 14 14.42 14.5 14.58 14.67 14.75 14.83 14.92 15 15.08 15.17 15.25
#>    13.92     1  0     0    0     0     0     0     0     0  0     0     0     0
#>    14        0  1     0    0     0     0     0     0     0  0     0     0     0
#>    14.42     0  0     1    0     0     0     0     0     0  0     0     0     0
#>    14.5      0  0     0    2     0     0     0     0     0  0     0     0     0
#>    14.58     0  0     0    0     8     0     1     0     0  0     0     0     0
#>    14.67     0  0     0    0     0     7     0     0     0  0     0     0     0
#>    14.75     0  0     0    0     0     0    10     0     0  0     0     1     0
#>    14.83     0  0     0    0     0     0     0    11     0  0     0     0     0
#>    14.92     0  0     0    0     0     0     0     0    27  0     0     0     0
#>    15        0  0     0    0     1     0     0     1     0 30     0     0     1
#>    15.08     0  0     0    0     0     0     0     1     1  0    21     0     0
#>    15.17     0  0     0    0     1     0     0     0     1  0     0    32     0
#>    15.25     0  0     0    0     0     0     0     0     0  0     0     0    17
#>    15.33     0  0     0    0     0     0     0     1     0  0     1     1     1
#>    15.42     0  0     0    0     0     0     0     0     0  1     0     0     0
#>    15.5      0  0     0    0     0     0     1     0     1  2     0     0     0
#>    15.58     0  0     0    0     0     0     0     2     0  0     0     1     0
#>    15.67     0  0     0    0     0     0     0     0     0  0     0     0     1
#>    15.75     0  0     0    0     0     0     0     0     0  1     1     0     0
#>    15.83     0  0     0    0     0     0     0     0     1  0     1     0     1
#>    15.92     0  0     0    0     0     0     0     0     0  1     0     1     0
#>    16        0  0     0    0     0     0     0     0     0  0     0     1     0
#>    16.08     0  0     0    0     0     0     0     0     0  0     0     0     0
#>    16.17     0  0     0    0     0     0     0     0     0  0     0     0     0
#>    16.25     0  0     0    0     0     0     0     0     0  1     0     0     1
#>    16.33     0  0     0    0     0     0     1     0     0  0     0     1     0
#>         GADSdat2
#> GADSdat1 15.33 15.42 15.5 15.58 15.67 15.75 15.83 15.92 16 16.08 16.17 16.25
#>    13.92     0     0    0     0     0     0     0     0  0     0     0     0
#>    14        0     0    0     0     0     0     0     0  0     0     0     0
#>    14.42     0     0    0     0     0     0     0     0  0     0     0     0
#>    14.5      0     0    0     0     0     0     0     0  0     0     0     0
#>    14.58     0     0    0     0     0     0     0     0  0     0     0     1
#>    14.67     0     0    0     0     0     0     0     0  0     0     0     0
#>    14.75     0     0    0     0     0     1     0     0  0     0     0     0
#>    14.83     1     1    0     0     0     1     0     0  1     0     0     0
#>    14.92     0     0    3     0     0     0     0     0  0     0     0     0
#>    15        1     0    0     0     2     0     0     0  0     0     0     0
#>    15.08     0     0    0     0     0     0     0     0  0     0     0     0
#>    15.17     1     0    1     0     1     0     1     0  0     0     0     0
#>    15.25     2     1    0     1     0     0     1     0  0     0     0     0
#>    15.33    20     0    1     1     1     0     1     0  0     1     0     0
#>    15.42     0    28    0     1     0     1     1     0  0     0     0     1
#>    15.5      0     0   35     2     0     1     0     0  1     0     0     0
#>    15.58     1     0    1    24     2     0     0     0  0     0     0     0
#>    15.67     0     1    0     0    25     1     0     0  1     0     1     1
#>    15.75     0     0    1     0     0    22     0     2  0     1     0     0
#>    15.83     0     0    0     2     0     0    23     0  0     0     0     0
#>    15.92     0     0    0     0     0     0     0    10  0     0     0     0
#>    16        0     1    0     0     0     1     0     0  7     0     0     0
#>    16.08     0     0    1     0     0     0     0     0  0     3     0     0
#>    16.17     1     0    0     0     0     0     0     0  0     0     7     0
#>    16.25     1     0    0     0     0     0     0     0  0     0     0     0
#>    16.33     1     1    0     0     0     0     1     0  0     0     0     0
#>         GADSdat2
#> GADSdat1 16.33
#>    13.92     0
#>    14        0
#>    14.42     0
#>    14.5      0
#>    14.58     0
#>    14.67     0
#>    14.75     1
#>    14.83     1
#>    14.92     1
#>    15        0
#>    15.08     1
#>    15.17     0
#>    15.25     0
#>    15.33     0
#>    15.42     0
#>    15.5      0
#>    15.58     0
#>    15.67     0
#>    15.75     0
#>    15.83     0
#>    15.92     0
#>    16        0
#>    16.08     1
#>    16.17     0
#>    16.25     0
#>    16.33    32
#> 
#> $some_unequals_GADSdat1
#>     idstud idschool idclass schtype sameteach g8g9 ganztag classsize   age
#> 400    402      133     407       1         1    1       1        22 15.00
#> 401    403       27      97       1         1    1       2        22 15.67
#> 402    404       35     119       2         2   NA       1        19 14.75
#> 403    405       74     237       3        NA   NA       1        19 15.75
#> 404    406      119     365       3         2   NA       1        20 15.50
#> 
#> $some_unequals_GADSdat2
#>     idstud idschool idclass schtype sameteach g8g9 ganztag classsize   age
#> 400    402      133     407       1         1    1       1        22 14.58
#> 401    403       27      97       1         1    1       2        22 16.25
#> 402    404       35     119       2         2   NA       1        19 16.33
#> 403    405       74     237       3        NA   NA       1        19 15.92
#> 404    406      119     365       3         2   NA       1        20 14.92
#> 
#> $unequal_IDs
#>  [1] 402 403 404 405 406 407 408 409 410 412 413 414 415 416 417 418 419 420 421
#> [20] 422 423 424 425 426 427 428 429 430 431 432 433 434 435 436 437 438 439 440
#> [39] 441 442 443 444 445 446 447 448 449 450 451 452 453 454 455 456 457 458 459
#> [58] 461 462 463 464 465 466 467 469 470 472 474 475 476 477 478 479 480 481 482
#> [77] 483 484 485 486 487 488 489 490 491 492 493 494 495 496 497 498 499 500 501
#> [96] 502
#>